+++ src/pipecat/services/aws/stt.py
@@ -284,7 +284,8 @@ class AWSTranscribeSTTService(STTService):
                 break
 
             try:
-                response = await self._ws_client.recv()
+                response = await asyncio.wait_for(self._ws_client.recv(), timeout=1.0)
+
                 headers, payload = decode_event(response)
 
                 if headers.get(":message-type") == "event":
@@ -334,6 +335,8 @@ class AWSTranscribeSTTService(STTService):
                 else:
                     logger.debug(f"{self} Other message type received: {headers}")
                     logger.debug(f"{self} Payload: {payload}")
+            except asyncio.TimeoutError:
+                self.reset_watchdog()
             except websockets.exceptions.ConnectionClosed as e:
                 logger.error(
                     f"{self} WebSocket connection closed in receive loop with code {e.code}: {e.reason}"